@@ -1,13 +1,9 @@
-const fs = require('fs').promises
-const path = require('path')
+import * as fs from 'fs/promises'
+import * as path from 'path'
 
-const log = async (...args) => {
+export const log = async (...args: any[]) => {
   const filename = path.resolve(__dirname, '../../../auguste.log')
   const timestamp = new Date().toISOString()
   const toWrite = JSON.stringify([timestamp, ...args], null, 2)
   await fs.appendFile(filename, toWrite)
-}
-
-module.exports = {
-  log,
 }
