@@ -1,21 +1,16 @@
-const path = require('path')
-const fs = require('fs').promises
-const { app } = require('electron')
+import * as path from 'path'
+import * as fs from 'fs/promises'
+import { app } from 'electron'
 
 const appPath = app.getPath('userData')
 
-const write = async (key, value) => {
+export const write = async (key: string, value: string) => {
   const filePath = path.resolve(appPath, key)
   await fs.writeFile(filePath, value)
 }
 
-const get = async key => {
+export const get = async (key: string) => {
   const filePath = path.resolve(appPath, key)
   const file = await fs.readFile(filePath, 'utf-8')
   return file
-}
-
-module.exports = {
-  write,
-  get,
 }
