@@ -5,13 +5,13 @@ const toClName = (...classes) => {
   return classes.filter(v => !!v).join(' ')
 }
 
-const Card = ({ tag = 'div', panel, children, ...props }) => {
+export const Card = ({ tag = 'div', panel, children, ...props }) => {
   const style = { gridArea: panel }
   const className = toClName(props.className, styles.card)
   return React.createElement(tag, { style, className }, children)
 }
 
-const Header = ({ title, ...props }) => {
+export const Header = ({ title, ...props }) => {
   const className = toClName(props.className, styles.cardHeader)
   return (
     <div className={className}>
@@ -21,12 +21,7 @@ const Header = ({ title, ...props }) => {
   )
 }
 
-const Body = ({ children, center }) => {
+export const Body = ({ children, center }) => {
   const cl = center ? styles.cardBodyCenter : styles.cardBody
   return <div className={cl}>{children}</div>
 }
-
-Card.Header = Header
-Card.Body = Body
-
-export default Card
