@@ -1,4 +1,4 @@
-const ComputerIcon = ({ color, ...props }) => (
+export const ComputerIcon = ({ color, ...props }) => (
   <svg width="93px" height="78px" viewBox="0 0 93 78" version="1.1" {...props}>
     <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
       <g
@@ -15,5 +15,3 @@ const ComputerIcon = ({ color, ...props }) => (
     </g>
   </svg>
 )
-
-export default ComputerIcon
